fix: do not delete events if ids is empty

+++ main/http-server/services/controlid-events.service.ts
@@ -296,9 +296,15 @@ async function create(event: ControlIdEvent): Promise<void> {
     socketServer.broadcast(message);
   }
 
-  const ids = event.object_changes.map((change) => change.values.id);
+  const ids = event.object_changes
+    .filter((change) => Boolean(change.values))
+    .map((change) => change.values.id);
 
   const client = await new ControlidClient().init(ip, equipamento.porta);
+
+  if (!ids.length) {
+    return;
+  }
 
   await client.deleteEvents({ ids });
 }
